Fixes + Beam search

+++ deep_learning/hw4p1/hw4lib/data/asr_dataset.py
@@ -248,7 +248,7 @@ class ASRDataset(Dataset):
                 - golden_transcript: LongTensor  (time) or None
         """
         # TODO: Load features
-        feat = torch.LongTensor(self.feats[idx])
+        feat = torch.FloatTensor(self.feats[idx])
 
         # TODO: Apply normalization
         if self.config['norm'] == 'global_mvn':
